Miscellanea: Methods for Fitted GVF Models

These methods extract information from fitted GVF model(s).

# S3 method for gvf.fit
coef(object, ...)
# S3 method for gvf.fits
coef(object, ...)
# S3 method for gvf.fit.gr
coef(object, ...)
# S3 method for gvf.fits.gr
coef(object, ...)

# S3 method for gvf.fit
residuals(object, ...)
# S3 method for gvf.fits
residuals(object, ...)
# S3 method for gvf.fit.gr
residuals(object, ...)
# S3 method for gvf.fits.gr
residuals(object, ...)

# S3 method for gvf.fit
fitted(object, ...)
# S3 method for gvf.fits
fitted(object, ...)
# S3 method for gvf.fit.gr
fitted(object, ...)
# S3 method for gvf.fits.gr
fitted(object, ...)

# S3 method for gvf.fit
predict(object, ...)
# S3 method for gvf.fits
predict(object, ...)
# S3 method for gvf.fit.gr
predict(object, ...)
# S3 method for gvf.fits.gr
predict(object, ...)

# S3 method for gvf.fit
effects(object, ...)
# S3 method for gvf.fits
effects(object, ...)
# S3 method for gvf.fit.gr
effects(object, ...)
# S3 method for gvf.fits.gr
effects(object, ...)

# S3 method for gvf.fit
rstandard(model, ...)
# S3 method for gvf.fits
rstandard(model, ...)
# S3 method for gvf.fit.gr
rstandard(model, ...)
# S3 method for gvf.fits.gr
rstandard(model, ...)

# S3 method for gvf.fit
rstudent(model, ...)
# S3 method for gvf.fits
rstudent(model, ...)
# S3 method for gvf.fit.gr
rstudent(model, ...)
# S3 method for gvf.fits.gr
rstudent(model, ...)

# S3 method for gvf.fit
anova(object, ...)
# S3 method for gvf.fits
anova(object, ...)
# S3 method for gvf.fit.gr
anova(object, ...)
# S3 method for gvf.fits.gr
anova(object, ...)

# S3 method for gvf.fit
vcov(object, ...)
# S3 method for gvf.fits
vcov(object, ...)
# S3 method for gvf.fit.gr
vcov(object, ...)
# S3 method for gvf.fits.gr
vcov(object, ...)

Arguments

object
An object containing one or more fitted GVF models (see ‘Usage’ for the allowed classes).

model
An object containing one or more fitted GVF models (see ‘Usage’ for the allowed classes).

...
Further arguments passed to or from other methods (see corresponding .lm methods).

Details

These methods can be used to extract information from fitted GVF model(s).

For more details on their usage, please read the help pages of the methods with same name defined on class .lm by package stats (e.g. coef, fitted, etc.).

Value

The requested information, wrapped into an R object whose structure depends on the class of the input fitted GVF model(s) (i.e. arguments object and/or model).

See also

Examples

# Load example data:
data(AF.gvf)

# Inspect available estimates and errors of counts:
head(ee.AF)

#>               name        Y        SE         CV       VAR
#> 1              ind 924101.3 17172.681 0.01858312 294900983
#> 2             sexf 470060.8 12796.386 0.02722283 163747492
#> 3             sexm 454040.5 11359.582 0.02501887 129040102
#> 4   marstatmarried 536680.2 13970.474 0.02603128 195174152
#> 5 marstatunmarried 312654.5 11180.844 0.03576102 125011282
#> 6   marstatwidowed  74766.6  5880.564 0.07865228  34581029

summary(ee.AF)

#>        name           Y                SE              CV         
#>  age10c1 :  1   Min.   :     0   Min.   :    0   Min.   :0.01858  
#>  age10c10:  1   1st Qu.:  4292   1st Qu.: 1212   1st Qu.:0.08230  
#>  age10c2 :  1   Median : 16633   Median : 2268   Median :0.14861  
#>  age10c3 :  1   Mean   : 45014   Mean   : 2989   Mean   :0.22191  
#>  age10c4 :  1   3rd Qu.: 46046   3rd Qu.: 4006   3rd Qu.:0.27813  
#>  age10c5 :  1   Max.   :924101   Max.   :17173   Max.   :1.00000  
#>  (Other) :343                                    NA's   :6        
#>       VAR           
#>  Min.   :        0  
#>  1st Qu.:  1470114  
#>  Median :  5146089  
#>  Mean   : 15409738  
#>  3rd Qu.: 16045307  
#>  Max.   :294900983  
#>                     

# List available registered GVF models:
GVF.db

#> 
#> # Registered GVF models currently available:
#> 
#>   Model.id                GVF.model Estimator.kind      Resp.to.CV
#> 1        1       log(CV^2) ~ log(Y)      Frequency sqrt(exp(resp))
#> 2        2            CV^2 ~ I(1/Y)      Frequency      sqrt(resp)
#> 3        3 CV^2 ~ I(1/Y) + I(1/Y^2)      Frequency      sqrt(resp)
#> 4        4          SE ~ Y + I(Y^2)          Total          resp/Y
#> 5        5          CV ~ I(1/Y) + Y          Total            resp
#> 

## (A) A *single* fitted GVF model ##
# Fit example data to registered GVF model number one:
m <- fit.gvf(ee.AF, 1)

# Extract some information:
coef(m)

#> (Intercept)      log(Y) 
#>   5.7518487  -0.9948979 

fitted(m)

#>           1           2           3           4           5           6 
#> -7.91464226 -7.24213145 -7.20763272 -7.37399555 -6.83644862 -5.41302091 
#>           7           8           9          10          11          12 
#> -5.95512677 -6.77718776 -6.96691378 -5.90746547 -4.16078349 -4.93077019 
#>          13          14          15          16          17          18 
#> -5.51554300 -5.98608773 -6.17966521 -6.36174913 -6.18501900 -5.64967262 
#>          19          20          21          22          23          24 
#> -4.43775737 -3.89374587 -2.72170942 -6.77340871 -7.10775313 -6.48663997 
#>          25          26          27          28          29          30 
#> -4.94776887 -5.39417556 -6.51911555 -5.24578784 -5.93722214 -4.59645882 
#>          31          32          33          34          35          36 
#> -5.20464333 -6.16052664 -5.21832910 -5.39312537 -6.70358729 -6.16458662 
#>          37          38          39          40          41          42 
#> -4.72252631 -6.66480459 -6.12876694 -4.72429344 -5.35306515 -6.08566265 
#>          43          44          45          46          47          48 
#> -6.27477769 -5.28566119 -3.31761501 -5.16951305 -6.08948790 -6.27982219 
#>          49          50          51          52          53          54 
#> -5.14508697 -3.60416817 -4.06492182 -6.27389141 -6.56979161 -5.53343247 
#>          55          56          57          58          59          60 
#> -3.89397255 -5.76924317 -5.66745967 -5.55009622 -4.19041938 -1.49266011 
#>          61          62          63          64          65          66 
#> -2.10385929 -4.12071873 -4.55325900 -3.91772470 -2.37890110 -6.08892547 
#>          67          68          69          70          71          72 
#> -6.45243670 -5.79748739 -6.07864410 -6.38262382 -5.79657106 -4.44095468 
#>          73          74          75          76          77          78 
#> -4.74499217 -5.85805912 -4.47921550 -5.25077981 -3.86806882 -4.51503269 
#>          79          80          81          82          83          84 
#> -5.47003679 -4.53261576 -4.73518331 -4.03404178 -4.66242507 -5.80010689 
#>          85          86          87          88          89          90 
#> -4.62761051 -5.24443307 -3.94417251 -4.51503269 -5.47179442 -4.52480583 
#>          91          92          93          94          95          96 
#> -4.67080486 -3.54072372 -5.62978583 -5.84203709 -4.96616194 -2.97153815 
#>          97          98          99         100         101         102 
#> -5.14166693 -4.92500186 -4.92882804 -3.35981806 -1.11257071 -1.85320792 
#>         103         104         105         106         107         108 
#> -3.22095258 -3.92457366 -3.25857264 -1.63769693 -3.17680320 -5.53659419 
#>         109         110         111         112         113         114 
#> -5.91691628 -4.70430043 -3.39290090 -5.01347139 -5.02803008 -4.78709991 
#>         115         116         117         118         119         120 
#> -3.62427665 -0.35133432 -0.60962000 -3.60452059 -3.79874733 -3.19669346 
#>         121         122         123         124         125         126 
#> -1.73833982 -2.53232948 -3.09215992 -4.84404227 -5.02783216 -5.21720861 
#>         127         128         129         130         131         132 
#> -5.08313040 -4.70017635 -3.52366318 -2.83335901 -0.93923107 -4.06794542 
#>         133         134         135         136         137         138 
#> -4.72848579 -4.13926428 -4.32304321 -4.37059429 -4.08781409 -3.05199982 
#>         139         140         141         142         143         144 
#> -2.04272892 -1.11257071          NA -1.47051037 -0.71746479 -2.52299309 
#>         145         146         147         148         149         150 
#> -2.53962130 -3.29333288 -3.17295402 -3.18121455 -0.67898118 -1.38145822 
#>         151         152         153         154         155         156 
#> -0.16271647 -2.12060472 -2.75447895 -4.66042538 -5.00402197 -5.34422806 
#>         157         158         159         160         161         162 
#> -5.09478929 -4.44632409 -3.23521248 -3.01401479 -2.24898394 -4.00554593 
#>         163         164         165         166         167         168 
#> -4.56464043 -4.25234263 -4.41763842 -4.11124083 -4.08354495 -3.26612007 
#>         169         170         171         172         173         174 
#> -2.43410771  0.26815625  0.41371570  0.26815625 -0.07818422 -3.13217222 
#>         175         176         177         178         179         180 
#> -2.63655039 -3.12971887 -3.08811970 -2.98083035 -1.57051306 -1.45456457 
#>         181         182         183         184         185         186 
#> -0.35176539 -5.57334670 -5.91912941 -5.21678081 -5.02983503 -5.30395829 
#>         187         188         189         190         191         192 
#> -4.82181737 -3.28203165 -4.15845237 -3.10973647 -5.54253725 -5.81852194 
#>         193         194         195         196         197         198 
#> -5.28230862 -5.03065811 -5.30194876 -4.67855800 -3.40149006 -4.04370095 
#>         199         200         201         202         203         204 
#> -3.26583217 -3.71627317 -4.21589684 -5.41469482 -3.70446326 -4.70820755 
#>         205         206         207         208         209         210 
#> -3.37298702 -4.10112971 -4.85977113 -4.02396005 -4.16856736 -3.31279258 
#>         211         212         213         214         215         216 
#> -3.65979318 -4.60025821 -3.59973171 -4.19257705 -2.77999153 -3.32909435 
#>         217         218         219         220         221         222 
#> -4.51635422 -3.49820422 -3.77673839 -2.81738149 -2.18762987 -3.30891774 
#>         223         224         225         226         227         228 
#> -2.43304424 -2.67589197 -1.01859902 -1.22771123 -2.89366595 -1.48441164 
#>         229         230         231         232         233         234 
#> -1.80508587 -3.49228690 -4.01076183 -5.25487524 -4.07588546 -4.69957138 
#>         235         236         237         238         239         240 
#> -3.41800438 -4.01299818 -4.93327770 -4.07075472 -4.11557933 -2.89245975 
#>         241         242         243         244         245         246 
#> -3.75064385 -4.70492964 -3.42888639 -4.21796906 -2.93939443 -3.49729193 
#>         247         248         249         250         251         252 
#> -4.41363511 -3.23258404 -3.53259125 -1.76716877 -2.15735239 -3.39672522 
#>         253         254         255         256         257         258 
#> -2.56624341 -2.52967776 -0.89116730 -1.22771123 -2.86485733 -2.16796436 
#>         259         260         261         262         263         264 
#> -2.46719214 -2.51897527 -2.64688131 -2.09869830 -4.40501556 -4.89742546 
#>         265         266         267         268         269         270 
#> -4.16346432 -4.78253618 -4.97434572 -4.41050174 -3.86674757 -4.22635978 
#>         271         272         273         274         275         276 
#> -3.33157711 -1.25346255 -2.59251166 -1.00807090 -3.94336258 -4.38076729 
#>         277         278         279         280         281         282 
#> -3.69968636 -3.85648297 -3.98626547 -3.61908968 -3.85338666 -3.97322746 
#>         283         284         285         286         287         288 
#> -3.65532599 -1.80513588 -2.65848549 -2.14956609 -0.57544980 -0.24279167 
#>         289         290         291         292         293         294 
#>          NA -1.09035840 -0.68068570 -0.38127588 -1.64318433 -2.54795084 
#>         295         296         297         298         299         300 
#> -1.97839366 -2.60509934 -3.30216686 -2.32128390 -1.63253602 -2.89200704 
#>         301         302         303         304         305         306 
#> -1.09035840          NA -1.49327600  0.35370140 -1.75409639 -2.54060066 
#>         307         308         309         310         311         312 
#> -1.72801932 -4.50483775 -4.66533528 -4.06982601 -4.77828644 -5.02110693 
#>         313         314         315         316         317         318 
#> -4.63324716 -3.47896137 -4.00956268 -3.15516928 -2.23229729 -2.56995928 
#>         319         320         321         322         323         324 
#> -2.01873510 -3.93344448 -4.07821445 -3.71699797 -3.92587374 -4.25318405 
#>         325         326         327         328         329         330 
#> -3.47898926 -3.78129731 -3.93611387 -3.23912441 -2.12318685 -3.06751409 
#>         331         332         333         334         335         336 
#> -2.05917130          NA          NA -0.35133432          NA -0.07818422 
#>         337         338         339         340         341         342 
#>  0.26815625 -1.87367339 -2.99334400 -2.35011928 -2.75569139 -2.95634310 
#>         343         344         345         346         347         348 
#> -2.29544408 -1.83900445 -2.65542694 -1.39883513 -0.50189202 -1.07360226 
#>         349 
#> -0.13094481 

## (B) A *a set* of GVF models fitted to the same data ##
# Fit example data to registered GVF models for frequencies (i.e. number 1:3):
mm <- fit.gvf(ee.AF, 1:3)

# Extract some information:
r.mod <- residuals(mm)
lapply(r.mod, head)

#> [[1]]
#>           1           2           3           4           5           6 
#> -0.05636149  0.03473291 -0.16861685  0.07708302  0.17465511  0.32758362 
#> 
#> [[2]]
#>           1           2           3           4           5           6 
#> -0.01311595 -0.01300715 -0.01314289 -0.01299811 -0.01276340 -0.01064978 
#> 
#> [[3]]
#>            1            2            3            4            5            6 
#> 0.0004512901 0.0004519346 0.0003084251 0.0004882947 0.0005849223 0.0016497883 
#> 

r.sta <- rstandard(mm)
lapply(r.sta, head)

#> [[1]]
#>           1           2           3           4           5           6 
#> -0.13777389  0.08467353 -0.41101035  0.18800930  0.42518703  0.79467755 
#> 
#> [[2]]
#>          1          2          3          4          5          6 
#> -0.2249911 -0.2231242 -0.2254528 -0.2229693 -0.2189426 -0.1826822 
#> 
#> [[3]]
#>           1           2           3           4           5           6 
#> 0.008803815 0.008816321 0.006016741 0.009525650 0.011410552 0.032181511 
#> 

r.stu <- rstudent(mm)
lapply(r.stu, head)

#> [[1]]
#>           1           2           3           4           5           6 
#> -0.13757555  0.08455018 -0.41050895  0.18774316  0.42467572  0.79424727 
#> 
#> [[2]]
#>          1          2          3          4          5          6 
#> -0.2246776 -0.2228131 -0.2251387 -0.2226583 -0.2186367 -0.1824230 
#> 
#> [[3]]
#>           1           2           3           4           5           6 
#> 0.008790860 0.008803347 0.006007887 0.009511633 0.011393762 0.032134199 
#> 

## (C) a *single* GVF model fitted to *grouped* data ##
# We have at our disposal the following survey design object on household data:
exdes

#> Stratified 2 - Stage Cluster Sampling Design (with replacement)
#> - [55] strata
#> - [1307, 2372] clusters
#> 
#> Call:
#> e.svydesign(data = example, ids = ~towcod + famcod, strata = ~SUPERSTRATUM, 
#>     weights = ~weight, fpc = NULL, self.rep.str = NULL, check.data = TRUE)

# Use function svystat to prepare *grouped* estimates and errors of counts
# to be fitted separately (here groups are regions):
ee.g <- svystat(exdes, y=~ind, by=~age5c:marstat:sex, combo=3, group=~regcod)
str(ee.g)

#> List of 3
#>  $ 6 :Classes ‘gvf.input’ and 'data.frame':	69 obs. of  5 variables:
#>   ..$ name: chr [1:69] "6:ind" "6.1:ind" "6.2:ind" "6.3:ind" ...
#>   ..$ Y   : num [1:69] 293458 40903 93388 118888 34964 ...
#>   ..$ SE  : num [1:69] 8924 4080 6249 5682 4605 ...
#>   ..$ CV  : num [1:69] 0.0304 0.0998 0.0669 0.0478 0.1317 ...
#>   ..$ VAR : num [1:69] 79634317 16648310 39046122 32287405 21209323 ...
#>   ..- attr(*, "y.vars")= chr "ind"
#>   ..- attr(*, "stats.kind")= chr "Absolute Frequency"
#>   ..- attr(*, "has.Deff")= logi FALSE
#>  $ 7 :Classes ‘gvf.input’ and 'data.frame':	71 obs. of  5 variables:
#>   ..$ name: chr [1:71] "7:ind" "7.1:ind" "7.2:ind" "7.3:ind" ...
#>   ..$ Y   : num [1:71] 410672 55834 131681 148350 63303 ...
#>   ..$ SE  : num [1:71] 12714 4814 6149 6490 5796 ...
#>   ..$ CV  : num [1:71] 0.031 0.0862 0.0467 0.0437 0.0916 ...
#>   ..$ VAR : num [1:71] 1.62e+08 2.32e+07 3.78e+07 4.21e+07 3.36e+07 ...
#>   ..- attr(*, "y.vars")= chr "ind"
#>   ..- attr(*, "stats.kind")= chr "Absolute Frequency"
#>   ..- attr(*, "has.Deff")= logi FALSE
#>  $ 10:Classes ‘gvf.input’ and 'data.frame':	71 obs. of  5 variables:
#>   ..$ name: chr [1:71] "10:ind" "10.1:ind" "10.2:ind" "10.3:ind" ...
#>   ..$ Y   : num [1:71] 219971 32191 69506 89226 24631 ...
#>   ..$ SE  : num [1:71] 7322 2678 4026 3620 3092 ...
#>   ..$ CV  : num [1:71] 0.0333 0.0832 0.0579 0.0406 0.1255 ...
#>   ..$ VAR : num [1:71] 53611574 7170100 16205547 13102640 9562313 ...
#>   ..- attr(*, "y.vars")= chr "ind"
#>   ..- attr(*, "stats.kind")= chr "Absolute Frequency"
#>   ..- attr(*, "has.Deff")= logi FALSE
#>  - attr(*, "group.vars")= chr "regcod"
#>  - attr(*, "class")= chr [1:2] "gvf.input.gr" "list"

# Fit registered GVF model number one separately inside groups:
m.g <- fit.gvf(ee.g, 1)

# Extract some information:
coef(m.g)

#> $`6`
#> (Intercept)      log(Y) 
#>   5.3879239  -0.9527893 
#> 
#> $`7`
#> (Intercept)      log(Y) 
#>    6.152759   -1.017113 
#> 
#> $`10`
#> (Intercept)      log(Y) 
#>    5.721254   -1.016518 
#> 

fitted(m.g)

#> $`6`
#>          1          2          3          4          5          6          7 
#> -6.6072082 -4.7297156 -5.5162894 -5.7463127 -4.5802393 -2.7852877 -6.1037229 
#>          8          9         10         11         12         13         14 
#> -5.5982478 -3.9829423 -5.9516953 -5.9418491 -2.8957376 -5.0484764 -5.3589888 
#>         15         16         17         18         19         20         21 
#> -4.3162872 -2.5609311 -4.5526282 -4.5079693 -4.4373039 -2.6740531 -0.6715748 
#>         22         23         24         25         26         27         28 
#> -1.1646902 -2.4712860 -3.3505789 -2.4483286 -0.6011303 -4.1437910 -4.8095031 
#>         29         30         31         32         33         34         35 
#> -5.0901459 -4.0221788 -1.7109306 -3.9884707 -4.9000776 -5.0816140 -3.8051175 
#>         36         37         38         39         40         41         42 
#> -2.4124761 -5.4579382 -4.9374304 -3.2636019 -5.4284327 -4.9382186 -3.3780043 
#>         43         44         45         46         47         48         49 
#> -2.5328415 -4.3390561 -4.7005983 -3.8235700 -1.3208910 -3.8969423 -3.8137399 
#>         50         51         52         53         54         55         56 
#> -3.8107746 -1.8492150 -0.6715748 -1.1646902 -1.6941180 -2.6153204 -1.6839204 
#>         57         58         59         60         61         62         63 
#> -1.8003358 -4.4346534 -4.6965285 -3.4521967 -2.2582971 -3.8874440 -3.8801937 
#>         64         65         66         67         68         69 
#> -3.7417364 -2.1538047 -1.9148517 -2.7595388 -1.8816501 -0.6011303 
#> 
#> $`7`
#>           1           2           3           4           5           6 
#> -6.99399032 -4.96442693 -5.83711622 -5.95834561 -5.09212083 -3.35771748 
#>           7           8           9          10          11          12 
#> -6.43370715 -5.85390036 -4.62687621 -6.32404106 -6.25266929 -3.08564238 
#>          13          14          15          16          17          18 
#> -5.32758352 -5.54213793 -4.64847258 -3.07147276 -4.76818014 -4.65337420 
#>          19          20          21          22          23          24 
#> -4.47568754 -3.38080340  0.02426166 -0.86637359 -3.29031945 -3.64694862 
#>          25          26          27          28          29          30 
#> -3.27535925 -1.76708909 -4.39095250 -5.14294763 -5.25989368 -4.43897109 
#>          31          32          33          34          35          36 
#> -2.73852031 -4.10831008 -5.12114954 -5.24673650 -4.33246521 -2.55898367 
#>          37          38          39          40          41          42 
#> -5.77882532 -5.14991780 -3.97883341 -5.67597136 -5.14786340 -3.86151966 
#>          43          44          45          46          47          48 
#> -2.43350987 -4.73430734 -4.81294517 -4.04825715 -2.37792617 -4.20611250 
#>          49          50          51          52          53          54 
#> -3.80280168 -3.78947255 -2.44537316  0.02426166 -0.42341027 -2.33237034 
#>          55          56          57          58          59          60 
#> -3.10342756 -2.68410911 -1.25414524 -2.32485603 -4.49703472 -4.86075053 
#>          61          62          63          64          65          66 
#> -3.82661911 -2.35487021 -3.89680383 -4.07568040 -3.75153022 -2.86353513 
#>          67          68          69          70          71 
#>  0.19254470 -2.78770886 -2.74988175 -2.44224632 -0.82510043 
#> 
#> $`10`
#>          1          2          3          4          5          6          7 
#> -6.7831910 -4.8296433 -5.6120747 -5.8659592 -4.5575219 -2.8107631 -6.2243600 
#>          8          9         10         11         12         13         14 
#> -5.7162326 -4.1203687 -6.0790623 -6.0781261 -2.8326613 -5.0674153 -5.4866815 
#>         15         16         17         18         19         20         21 
#> -4.1780306 -2.5323222 -4.6491527 -4.4888673 -4.4043996 -3.0113331 -0.5145579 
#>         22         23         24         25         26         27         28 
#> -0.9711817 -3.0891098 -3.2187986 -2.1439956 -0.7762453 -4.1507637 -4.9428079 
#>         29         30         31         32         33         34         35 
#> -5.1336983 -3.9087738 -1.4159864 -4.0986620 -4.8708758 -5.1882937 -3.7938290 
#>         36         37         38         39         40         41         42 
#> -2.5134777 -5.4857363 -5.0821898 -3.3329034 -5.5526881 -4.9358173 -3.4923913 
#>         43         44         45         46         47         48         49 
#> -2.2998942 -4.4095300 -4.6619358 -3.5595649 -1.1855662 -3.9356736 -3.8533255 
#>         50         51         52         53         54         55         56 
#> -3.8903492 -2.3518674 -0.5451502 -2.1769752 -2.5273168 -1.2696419  0.2057991 
#>         57         58         59         60         61         62         63 
#> -1.9211599 -4.3138568 -4.8895218 -3.3793235 -2.2181933 -3.9533614 -3.7101805 
#>         64         65         66         67         68         69         70 
#> -3.4651031 -2.2595082 -0.5145579  0.1183949 -2.5567788 -2.5009155 -1.5848222 
#>         71 
#> -0.2893791 
#> 

## (D) a *set of* GVF models fitted to *grouped* data ##
# Fit all registered GVF models for frequencies (i.e. number 1:3) separately
# inside groups:
mm.g <- fit.gvf(ee.g, 1:3)
# Extract some information:
coef(mm.g)

#> $`6`
#> $`6`[[1]]
#> (Intercept)      log(Y) 
#>   5.3879239  -0.9527893 
#> 
#> $`6`[[2]]
#>  (Intercept)       I(1/Y) 
#> 9.833598e-03 2.709174e+02 
#> 
#> $`6`[[3]]
#>   (Intercept)        I(1/Y)      I(1/Y^2) 
#> -7.947002e-03  4.457342e+02 -1.071654e+05 
#> 
#> 
#> $`7`
#> $`7`[[1]]
#> (Intercept)      log(Y) 
#>    6.152759   -1.017113 
#> 
#> $`7`[[2]]
#>  (Intercept)       I(1/Y) 
#> 4.293585e-03 3.815012e+02 
#> 
#> $`7`[[3]]
#>   (Intercept)        I(1/Y)      I(1/Y^2) 
#>  1.325770e-03  4.162739e+02 -1.453588e+04 
#> 
#> 
#> $`10`
#> $`10`[[1]]
#> (Intercept)      log(Y) 
#>    5.721254   -1.016518 
#> 
#> $`10`[[2]]
#>  (Intercept)       I(1/Y) 
#> 5.101619e-03 2.552963e+02 
#> 
#> $`10`[[3]]
#>   (Intercept)        I(1/Y)      I(1/Y^2) 
#> -3.756025e-03  3.062188e+02 -1.480137e+04 
#> 
#> 

fitted(mm.g)

#> $`6`
#> $`6`[[1]]
#>          1          2          3          4          5          6          7 
#> -6.6072082 -4.7297156 -5.5162894 -5.7463127 -4.5802393 -2.7852877 -6.1037229 
#>          8          9         10         11         12         13         14 
#> -5.5982478 -3.9829423 -5.9516953 -5.9418491 -2.8957376 -5.0484764 -5.3589888 
#>         15         16         17         18         19         20         21 
#> -4.3162872 -2.5609311 -4.5526282 -4.5079693 -4.4373039 -2.6740531 -0.6715748 
#>         22         23         24         25         26         27         28 
#> -1.1646902 -2.4712860 -3.3505789 -2.4483286 -0.6011303 -4.1437910 -4.8095031 
#>         29         30         31         32         33         34         35 
#> -5.0901459 -4.0221788 -1.7109306 -3.9884707 -4.9000776 -5.0816140 -3.8051175 
#>         36         37         38         39         40         41         42 
#> -2.4124761 -5.4579382 -4.9374304 -3.2636019 -5.4284327 -4.9382186 -3.3780043 
#>         43         44         45         46         47         48         49 
#> -2.5328415 -4.3390561 -4.7005983 -3.8235700 -1.3208910 -3.8969423 -3.8137399 
#>         50         51         52         53         54         55         56 
#> -3.8107746 -1.8492150 -0.6715748 -1.1646902 -1.6941180 -2.6153204 -1.6839204 
#>         57         58         59         60         61         62         63 
#> -1.8003358 -4.4346534 -4.6965285 -3.4521967 -2.2582971 -3.8874440 -3.8801937 
#>         64         65         66         67         68         69 
#> -3.7417364 -2.1538047 -1.9148517 -2.7595388 -1.8816501 -0.6011303 
#> 
#> $`6`[[2]]
#>          1          2          3          4          5          6          7 
#> 0.01075679 0.01645694 0.01273459 0.01211236 0.01758198 0.06081062 0.01139958 
#>          8          9         10         11         12         13         14 
#> 0.01249548 0.02433690 0.01167048 0.01168957 0.05523088 0.01457364 0.01325533 
#>         15         16         17         18         19         20         21 
#> 0.02005531 0.07434541 0.01780980 0.01819256 0.01883609 0.06712332 0.47846770 
#>         22         23         24         25         26         27         28 
#> 0.28912986 0.08070985 0.03799841 0.08243835 0.51442919 0.02208399 0.01592489 
#>         29         30         31         32         33         34         35 
#> 0.01437080 0.02375178 0.16726132 0.02425299 0.01537251 0.01441161 0.02731281 
#>         36         37         38         39         40         41         42 
#> 0.08522245 0.01291781 0.01515956 0.04069049 0.01301481 0.01515516 0.03719927 
#>         43         44         45         46         47         48         49 
#> 0.07627563 0.01981394 0.01666248 0.02697755 0.24689815 0.02570689 0.02715535 
#>         50         51         52         53         54         55         56 
#> 0.02720934 0.14599351 0.47846770 0.28912986 0.17006389 0.07076594 0.17178803 
#>         57         58         59         60         61         62         63 
#> 0.15316096 0.01886117 0.01669171 0.03514919 0.09846427 0.02586592 0.02598838 
#>         64         65         66         67         68         69 
#> 0.02851511 0.10873743 0.13692938 0.06220705 0.14143632 0.51442919 
#> 
#> $`6`[[3]]
#>             1             2             3             4             5 
#> -0.0064293453  0.0028861870 -0.0031863565 -0.0042053928  0.0047135683 
#>             6             7             8             9            10 
#>  0.0721300539 -0.0053741127 -0.0035778167  0.0156078322 -0.0049297416 
#>            11            12            13            14            15 
#> -0.0048984510  0.0637349964 -0.0001811293 -0.0023343984  0.0087180000 
#>            16            17            18            19            20 
#>  0.0921162488  0.0050831762  0.0057037927  0.0067462604  0.0815182921 
#>            21            22            23            24            25 
#>  0.4424231347  0.3376760966  0.1013293885  0.0372337052  0.1038111423 
#>            26            27            28            29            30 
#>  0.4504882133  0.0119891684  0.0020206824 -0.0005121004  0.0146694219 
#>            31            32            33            34            35 
#>  0.2148790410  0.0154733226  0.0011212420 -0.0004454969  0.0203650662 
#>            36            37            38            39            40 
#>  0.1077901551 -0.0028865148  0.0007742681  0.0414309258 -0.0027278032 
#>            41            42            43            44            45 
#>  0.0007670902  0.0359836779  0.0949229176  0.0083279927  0.0032203103 
#>            46            47            48            49            50 
#>  0.0198304185  0.3000332151  0.0178010758  0.0201139916  0.0202000891 
#>            51            52            53            54            55 
#>  0.1890043684  0.4424231347  0.3376760966  0.2181901764  0.0868826362 
#>            56            57            58            59            60 
#>  0.2202157974  0.1978719323  0.0067868614  0.0032678211  0.0327684307 
#>            61            62            63            64            65 
#>  0.1264054194  0.0180553184  0.0182510510  0.0222796915  0.1404946393 
#>            66            67            68            69 
#>  0.1775754192  0.0742168427  0.1832882084  0.4504882133 
#> 
#> 
#> $`7`
#> $`7`[[1]]
#>           1           2           3           4           5           6 
#> -6.99399032 -4.96442693 -5.83711622 -5.95834561 -5.09212083 -3.35771748 
#>           7           8           9          10          11          12 
#> -6.43370715 -5.85390036 -4.62687621 -6.32404106 -6.25266929 -3.08564238 
#>          13          14          15          16          17          18 
#> -5.32758352 -5.54213793 -4.64847258 -3.07147276 -4.76818014 -4.65337420 
#>          19          20          21          22          23          24 
#> -4.47568754 -3.38080340  0.02426166 -0.86637359 -3.29031945 -3.64694862 
#>          25          26          27          28          29          30 
#> -3.27535925 -1.76708909 -4.39095250 -5.14294763 -5.25989368 -4.43897109 
#>          31          32          33          34          35          36 
#> -2.73852031 -4.10831008 -5.12114954 -5.24673650 -4.33246521 -2.55898367 
#>          37          38          39          40          41          42 
#> -5.77882532 -5.14991780 -3.97883341 -5.67597136 -5.14786340 -3.86151966 
#>          43          44          45          46          47          48 
#> -2.43350987 -4.73430734 -4.81294517 -4.04825715 -2.37792617 -4.20611250 
#>          49          50          51          52          53          54 
#> -3.80280168 -3.78947255 -2.44537316  0.02426166 -0.42341027 -2.33237034 
#>          55          56          57          58          59          60 
#> -3.10342756 -2.68410911 -1.25414524 -2.32485603 -4.49703472 -4.86075053 
#>          61          62          63          64          65          66 
#> -3.82661911 -2.35487021 -3.89680383 -4.07568040 -3.75153022 -2.86353513 
#>          67          68          69          70          71 
#>  0.19254470 -2.78770886 -2.74988175 -2.44224632 -0.82510043 
#> 
#> $`7`[[2]]
#>           1           2           3           4           5           6 
#> 0.005222553 0.011126385 0.007190738 0.006865212 0.010320222 0.037455775 
#>           7           8           9          10          11          12 
#> 0.005905101 0.007143322 0.013815593 0.006088570 0.006219050 0.047626292 
#>          13          14          15          16          17          18 
#> 0.009074769 0.008165479 0.013615544 0.048234194 0.012580513 0.013570728 
#>          19          20          21          22          23          24 
#> 0.015341598 0.036711555 0.926239339 0.388368037 0.039727677 0.029247873 
#>          25          26          27          28          29          30 
#> 0.040252711 0.162717717 0.016301427 0.010026462 0.009403788 0.015747703 
#>          31          32          33          34          35          36 
#> 0.065251842 0.020147981 0.010150651 0.009470322 0.017012156 0.077020038 
#>          37          38          39          40          41          42 
#> 0.007361624 0.009987309 0.022300304 0.007688104 0.009998821 0.024501713 
#>          43          44          45          46          47          48 
#> 0.086568629 0.012861138 0.012223700 0.021112251 0.091189960 0.018694474 
#>          49          50          51          52          53          54 
#> 0.025702660 0.025985070 0.085614574 0.926239339 0.597977297 0.095170468 
#>          55          56          57          58          59          60 
#> 0.046875166 0.068601648 0.266619895 0.095844341 0.015112139 0.011859600 
#>          61          62          63          64          65          66 
#> 0.025207155 0.093182225 0.023812703 0.020664847 0.026809529 0.058201533 
#>          67          68          69          70          71 
#> 1.092121224 0.062373992 0.064574710 0.085864958 0.404273825 
#> 
#> $`7`[[3]]
#>           1           2           3           4           5           6 
#> 0.002339325 0.008776698 0.004486152 0.004131134 0.007898090 0.037400769 
#>           7           8           9          10          11          12 
#> 0.003083912 0.004434442 0.011706628 0.003284041 0.003426365 0.048420600 
#>          13          14          15          16          17          18 
#> 0.006540462 0.005549079 0.011488721 0.049078611 0.010361170 0.011439903 
#>          19          20          21          22          23          24 
#> 0.013368589 0.036593590 0.922413577 0.405674914 0.039864184 0.028492380 
#>          25          26          27          28          29          30 
#> 0.040433329 0.171683185 0.014413694 0.007577900 0.006899146 0.013810796 
#>          31          32          33          34          35          36 
#> 0.067469084 0.018600146 0.007713265 0.006971675 0.015187448 0.080152795 
#>          37          38          39          40          41          42 
#> 0.004672512 0.007535224 0.020941369 0.005028539 0.007547772 0.023335028 
#>          43          44          45          46          47          48 
#> 0.090423898 0.010666901 0.009972412 0.019649160 0.095388369 0.017018548 
#>          49          50          51          52          53          54 
#> 0.024640447 0.024947381 0.089398470 0.922413577 0.613920716 0.099661016 
#>          55          56          57          58          59          60 
#> 0.047607455 0.071082308 0.280689608 0.100384033 0.013118716 0.009575689 
#>          61          62          63          64          65          66 
#> 0.024101872 0.097527247 0.022585950 0.019162460 0.025843348 0.059857040 
#>          67          68          69          70          71 
#> 1.070118853 0.064363141 0.066738432 0.089667603 0.421784949 
#> 
#> 
#> $`10`
#> $`10`[[1]]
#>          1          2          3          4          5          6          7 
#> -6.7831910 -4.8296433 -5.6120747 -5.8659592 -4.5575219 -2.8107631 -6.2243600 
#>          8          9         10         11         12         13         14 
#> -5.7162326 -4.1203687 -6.0790623 -6.0781261 -2.8326613 -5.0674153 -5.4866815 
#>         15         16         17         18         19         20         21 
#> -4.1780306 -2.5323222 -4.6491527 -4.4888673 -4.4043996 -3.0113331 -0.5145579 
#>         22         23         24         25         26         27         28 
#> -0.9711817 -3.0891098 -3.2187986 -2.1439956 -0.7762453 -4.1507637 -4.9428079 
#>         29         30         31         32         33         34         35 
#> -5.1336983 -3.9087738 -1.4159864 -4.0986620 -4.8708758 -5.1882937 -3.7938290 
#>         36         37         38         39         40         41         42 
#> -2.5134777 -5.4857363 -5.0821898 -3.3329034 -5.5526881 -4.9358173 -3.4923913 
#>         43         44         45         46         47         48         49 
#> -2.2998942 -4.4095300 -4.6619358 -3.5595649 -1.1855662 -3.9356736 -3.8533255 
#>         50         51         52         53         54         55         56 
#> -3.8903492 -2.3518674 -0.5451502 -2.1769752 -2.5273168 -1.2696419  0.2057991 
#>         57         58         59         60         61         62         63 
#> -1.9211599 -4.3138568 -4.8895218 -3.3793235 -2.2181933 -3.9533614 -3.7101805 
#>         64         65         66         67         68         69         70 
#> -3.4651031 -2.2595082 -0.5145579  0.1183949 -2.5567788 -2.5009155 -1.5848222 
#>         71 
#> -0.2893791 
#> 
#> $`10`[[2]]
#>           1           2           3           4           5           6 
#> 0.006262210 0.013032243 0.008774641 0.007962861 0.015466583 0.062891036 
#>           7           8           9          10          11          12 
#> 0.007112713 0.008416923 0.021036027 0.007421731 0.007423869 0.061659431 
#>          13          14          15          16          17          18 
#> 0.011378188 0.009256860 0.020157308 0.081100758 0.014573140 0.016190804 
#>          19          20          21          22          23          24 
#> 0.017151628 0.052543001 0.558289715 0.358110892 0.049048515 0.043784636 
#>          25          26          27          28          29          30 
#> 0.116458291 0.432733623 0.020566625 0.012196731 0.010981977 0.024723345 
#>          31          32          33          34          35          36 
#> 0.233004053 0.021379949 0.012716994 0.010674485 0.027072420 0.082522789 
#>          37          38          39          40          41          42 
#> 0.009260725 0.011287621 0.039677280 0.008995617 0.012245692 0.034656644 
#>          43          44          45          46          47          48 
#> 0.100625194 0.017090964 0.014454777 0.032766719 0.290987740 0.024210910 
#>          49          50          51          52          53          54 
#> 0.025823386 0.025082234 0.095863958 0.541889479 0.112903455 0.081475899 
#>          55          56          57          58          59          60 
#> 0.268293687 1.128764938 0.143751653 0.018274192 0.012578579 0.038133864 
#>          61          62          63          64          65          66 
#> 0.108619706 0.023881277 0.028956851 0.035460786 0.104496716 0.558289715 
#>          67          68          69          70          71 
#> 1.036185246 0.079294094 0.083485502 0.198127709 0.695464805 
#> 
#> $`10`[[3]]
#>             1             2             3             4             5 
#> -0.0023642444  0.0057421933  0.0006465699 -0.0003259265  0.0086519824 
#>             6             7             8             9            10 
#>  0.0648019067 -0.0013447080  0.0002180680  0.0152990696 -0.0009743558 
#>            11            12            13            14            15 
#> -0.0009717937  0.0633566226  0.0037635497  0.0012241169  0.0142512615 
#>            16            17            18            19            20 
#>  0.0860905533  0.0075843539  0.0095171308  0.0106645569  0.0526371003 
#>            21            22            23            24            25 
#>  0.5902776349  0.3913661363  0.0485181135  0.0423030550  0.1269962625 
#>            26            27            28            29            30 
#>  0.4676441895  0.0147393848  0.0047428768  0.0032894018  0.0196921016 
#>            31            32            33            34            35 
#>  0.2578094942  0.0157090759  0.0053651761  0.0029213758  0.0224875452 
#>            36            37            38            39            40 
#>  0.0877466833  0.0012287459  0.0036551742  0.0374447610  0.0009112427 
#>            41            42            43            44            45 
#>  0.0048014459  0.0314958071  0.1087488966  0.0105921234  0.0074428880 
#>            46            47            48            49            50 
#>  0.0292534676  0.3205933157  0.0190819621  0.0210014832  0.0201193485 
#>            51            52            53            54            55 
#>  0.1032393876  0.5746657038  0.1229092879  0.0865275411  0.2962024280 
#>            56            57            58            59            60 
#>  1.0573001480  0.1581840723  0.0120046018  0.0051996264  0.0356171848 
#>            61            62            63            64            65 
#>  0.1179766625  0.0186894154  0.0247282426  0.0324494052  0.1132212743 
#>            66            67            68            69            70 
#>  0.5902776349  0.9915568842  0.0839851465  0.0888673600  0.2193104564 
#>            71 
#>  0.7160749720 
#> 
#>